### 2023-04-11 #5316 @justaPCWingo
Hi @kthyng - Apologies. I replied to the email message, but failed to notice that it was from an issues thread.
- The comment at the head of `paper.md` has been removed from the document. It worked for me just prior to submission; perhaps I was using an older docker image?
- The user-doc folder is the source code for the user documentation, but it can be built quite easily with Sphinx by running `make.bat html` or `make.sh html` in the `user-doc` folder. The user documentation is included as part of the PyInstaller bundle distribution of this product, which I last checked is still going through the approval process on [EDX](https://edx.netl.doe.gov/).
