### 2023-04-11 #5316 @kthyng
> The user-doc folder is the source code for the user documentation, but it can be built quite easily with Sphinx by running make.bat html or make.sh html in the user-doc folder. The user documentation is included as part of the PyInstaller bundle distribution of this product, which I last checked is still going through the approval process on [EDX](https://edx.netl.doe.gov/).

Usually a user doesn't have to compile the docs themselves though right? Or you expect that most users would use the bundle? You might consider something like ReadTheDocs to host your docs.
