### 2023-04-11 #5316 @kthyng
I see this in the error for building your paper:

> Problem with ORCID (Author Without ORCID) for Devin Justman. ORCID looks malformed (Theoj::Error)
