### 2023-04-11 #5316 @justaPCWingo
Hi @kthyng I have removed the orcid entry for Devin Justman, as he doesn't have one. That was a placeholder that I mistakenly left in. Still not sure why this wasn't flagged when I ran the docker container supplied by the [JOSS website](https://joss.readthedocs.io/en/latest/submitting.html#docker).

Our lab ([NETL](https://netl.doe.gov/)) is mostly concerned with product releases, and is in the process of understanding the licensing and releasing of Open Source code. So to answer your question, the expectation is that most people will be using the PyInstaller bundled version, with the documentation compiled.

I would be fine releasing the documentation on ReadTheDocs, but unfortunately that's not up to me. Anything published by our lab has to go through a Rights In Data (RID) process before it can be approved to go public. I can look into what would be involved to get approval, but it would take at minimum 30 days to go through the process.

However, we are already using the RTD theme for our documentation, so it would be a straight port.
