### 2023-04-12 #5316 @justaPCWingo
HI @kthyng , Just a followup to the comment above: I spoke with folks on my side and it seems like we can route through the existing approval process for the bundled version of the tool. This means I can have a RTD submission ready to go within a few days. I'll go ahead and get started on submitting that to both RTD and our process, and update here once its posted. Thanks for the suggestion!
