+++ .readthedocs.yaml
@@ -1,0 +1,13 @@
+version: 1
+
+build:
+  os: "ubuntu-20.04"
+  tools:
+    python: "3.8"
+
+sphinx:
+  fail_on_warning: true
+  
+python:
+  install:
+    - requirements: user_doc/sphinx_requirements.txt
