### 2023-04-18 #5316 @justaPCWingo
Hi @kthyng , I've added the user documentation for the [URC Assessment Method to Read the Docs](https://urc-assessment-method.readthedocs.io/en/latest/).

The PyInstaller bundled version of this software can be found on [EDX](https://edx.netl.doe.gov/dataset/urc-assessment-method).

Please let me know what other issues you encounter with this submission. I'll endeavor to address them as quickly as possible.
