### 2023-05-09 #5316 @elbeejay
@justaPCWingo I've had a look and have the following questions and suggestions:

- Consider setting the default branch back to "master", we can track the paper in the "joss-submit" branch for this process, but it looks like you plan to continue development on "master" so you don't want folks opening up the repository and looking at any stale code in "joss-submit"
- While not a requirement for JOSS, it looks like there is `sphinx` documentation already written for this package. Have you considered building it and hosting it online somewhere (e.g., GitHub Pages)? That'd be significantly more accessible to users, developers, and reviewers, than asking them to build it locally
- Line 20 of the PDF please define the acronym "REE" before using it in the text
- Can you recommend any reviewers you think might be a good fit for this work? There is a [reviewer application(https://reviewers.joss.theoj.org/) with volunteer reviewers, as well as a [tool](https://github.com/dfm/joss-reviewer) that might help.
