### 2023-05-10 #5316 @justaPCWingo
Hi @elbeejay ,
  Below are my responses to the bullets you provided above.
   * I can make this change; I must have unintentionally marked it as default. The joss-submit branch was created precisely for this reason.
   * I have! based on previous discussions in this thread I have posted the [URC Assessment method on ReadTheDocs.io](https://urc-assessment-method.readthedocs.io/en/latest/).
   * Ack! Good catch; I see what happened here; REE stands for Rare-Earth Elements, and are defined as part of URC; in a previous revision that acronym was REE-URC. I'll update line 20 to define that acronym.
   * Thanks for the link to the reviewer search tool; I'll start with that. One of my colleagues suggested a form mutual colleague who used to be with our lab, but has since moved on to the Department of Defense. I may reach out to him if needed.

   I'll followup once I've carried out the tasks mentioned above.
   Thanks!
