### 2023-05-10 57bdd3c
Defined REE acronym in paper.md
