### 2023-05-10 #5316 @justaPCWingo
Quick update:
* I've updated the default branch, as suggested by @elbeejay .
* I've defined REE in the paper.md file as recommended by @elbeejay
* The joss-reviewer script is crashing when configured according to the readme. I've submitted an issue outlining the conditions  and output from the crash. I'm still contemplating if we have anyone we can recommend on our end.

Thanks.
