### 2023-05-22 #5316 @justaPCWingo
hi @elbeejay and @kthyng , I've found a researcher who is willing to be a review for my submission. Should I direct them to the JOSS website to register as a reviewer, or is there another method that you prefer when inviting reviewers?
