### 2023-05-22 #5316 @elbeejay
If they have a GitHub account we can invite them here in the issue. The only requirement JOSS has is a GitHub account, one does not ned to be registered on the JOSS webpage to conduct a review.
