### 2023-05-25 #5316 @justaPCWingo
Hi @elbeejay I missed your previous comments about Linking to the readthedocs in the readme, and requesting an example/test dataset. I've updated the readme to point to the documentation.

There are sample data provided as part of [this EDX submission](https://edx.netl.doe.gov/dataset/urc-assessment-method-publication-supplementary-files); specifically **esm_3.zip**. However, its not immediately obvious where/how to incorporate this data, so I'm writing a quick walkthrough/tutorial that will go through a complete run of the data with the tool. I hope to have it integrated with the user documentation by early next week.
