### 2023-05-26 #5316 @jeinsle
Hi Jay

Happy to have a read go at reviewing. my firsttime doing a code review, but i am working with a CompSci PhD on a project that i can being in on this as well.  He is building me a GUI so right up his alley.

Cheers

josh

----------------------------------------------------------------------------------------------------
Sometimes I work outside of standard hours, but I do not expect replies during these times.

Joshua F. Einsle
Lord Kelvin Adam Smith Research Fellow in Data Science
School of Geographical and Earth Sciences
University of Glasgow

https://www.gla.ac.uk/schools/ges/staff/joshuafranzeinsle/
