### 2023-05-26 #5316 @justaPCWingo
quick update: I've added a quick example walkthrough to the documentation which can be found [here](https://urc-assessment-method.readthedocs.io/en/latest/example.html).

Presently the CI action for testing is failing, but it appears to be from a bad GDAL installation in the ubuntu image rather than an issue with the code itself. I've had to change the .yml in the past to address this, but it seems to periodically crop up due to something outside of our control. I'm hoping to have that addressed soon as well.
