### 2023-05-26 #5316 @justaPCWingo
Quick followup: Fixed the CI testing by locking PIP to a specific version; I think the enforcement of some manifest rules broke how GDAL finds numpy, which caused my tests to fail!

Also thanks again @jeinsle for volunteering to review!
