### 2023-05-30 #5500 @editorialbot
[REVIEW]: A Python Tool for Predicting and Assessing Unconventional Rare-Earth and Critical Mineral Resources
**Submitting author:** @justaPCWingo (<a href="http://orcid.org/0000-0003-3934-7733">Patrick Wingo</a>)
**Repository:** https://github.com/NETL-RIC/URC-Assessment-Method
**Branch with paper.md** (empty if default branch): joss-submit
**Version:** v1.0.1
