### 2023-05-30 #5316 @jeinsle
Hi Jay

as noted my post-doc Francesco Perrone will be helping on my end.  His GitHub info is: FrancescoPerrone

We will get started on the checklist from tomorrow.

Cheers

josh

----------------------------------------------------------------------------------------------------
Sometimes I work outside of standard hours, but I do not expect replies during these times.

Joshua F. Einsle
Lord Kelvin Adam Smith Research Fellow in Data Science
School of Geographical and Earth Sciences
University of Glasgow

https://www.gla.ac.uk/schools/ges/staff/joshuafranzeinsle/
