### 2023-05-30 #5500 @jeinsle
for him it may be better for profile to be separate, but not clear how you want to handle.

----------------------------------------------------------------------------------------------------
Sometimes I work outside of standard hours, but I do not expect replies during these times.

Joshua F. Einsle
Lord Kelvin Adam Smith Research Fellow in Data Science
School of Geographical and Earth Sciences
University of Glasgow

https://www.gla.ac.uk/schools/ges/staff/joshuafranzeinsle/
