### 2023-06-01 #5500 @FrancescoPerrone
Hi there, apologies for not getting back sooner than now. I’m just acknowledging your email; I’ll be in touch if any additional question/help will be required.

Thanks a lot!
With kind regards,
Francesco

Sent from Outlook for iOS<https://aka.ms/o0ukef>
