### 2023-06-06 #5500 @elbeejay
@jameshgrn I don't know what happened there, maybe the bot was down for a bit. I'd say try once more to get it to generate checklist, and if that is unsuccessful we can find you the raw markdown to create a comment with your reviewer checklist manually.

Thanks!
