### 2023-06-07 #5500 @jameshgrn
Adding my thoughts in here. I agree with everything @FrancescoPerrone said so consider that a +1 for their recommendations

I tried to solve a conda environment for this package this morning and gave up after a little while of no success. Once an environment management system is implemented (requirements.txt or preferably an environment.yml file for conda) I will test out the package. below are some comments for the repository in general and the paper itself.

REPOSITORY COMMENTS
- Please include detailed installation instructions and a requirements.txt or an environment.yml file. This is crucial. I realize it is also in the readme, but id rather have pip or conda build the env for me as opposed to building it myself *especially* when GDAL is in play.
- Please include instructions to perform tests or implement a formal system for automated testing as discussed by @FrancescoPerrone
- The authors are encouraged to rename the COPYING file to 'LICENSE' to make it easier to find.
- "Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?" I dont see the API reference in the documentation. This should be easy to compile with sphinx and function docstrings..

PAPER COMMENTS
- Authors are encouraged to add detail to figure captions, hard to understand with the current captions.
- I think the paragraph that outlines the use of the equations needs to be expanded. It is currently hard to follow and some of the details are glossed over. If it is detailed in the sister paper to this one, just refer to that paper and dont include them.
- Authors are encouraged to add more to their manuscript, particularly a section after the statement of need. Maybe an example and more detailed description of what the package does / use cases, and where it fits into the open source landscape.
