### 2023-06-07 #5500 @justaPCWingo
@FrancescoPerrone and @jameshgrn Thanks for the feedback and comments. The requests seem reasonable to me. I'll dig in a bit and speak with my fellow authors.

@jameshgrn I hear you when it comes to GDAL; getting it installed is the bane of my existence. I can't tell you how many versions, platforms, and build configurations I've wrestled with over the years; its one of the biggest reason I push for PyInstaller bundles for distribution. I'll see what I can do with environment/requirement files. I should probably look into a proper pypi installation if time permits.

@elbeejay I'll shoot to address the Reviewer comments over the next two weeks. I have some other work obligations I need to balance with this, but I think that's a reasonable timeframe.
