### 2023-06-19 732a909
Added several components based on reviewer recommendations: environment recipes; expanded readme; version number and API reference in documentation
