+++ user_doc/source/api/urclib.rst
@@ -1,0 +1,94 @@
+urclib package
+==============
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   urclib.fuzzylogic
+   urclib.ui_qt
+
+Submodules
+----------
+
+urclib.calculate\_pe\_score module
+----------------------------------
+
+.. automodule:: urclib.calculate_pe_score
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.common\_utils module
+---------------------------
+
+.. automodule:: urclib.common_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.create\_pe\_grid module
+------------------------------
+
+.. automodule:: urclib.create_pe_grid
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.da\_calc module
+----------------------
+
+.. automodule:: urclib.da_calc
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.ds\_calc module
+----------------------
+
+.. automodule:: urclib.ds_calc
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.settings module
+----------------------
+
+.. automodule:: urclib.settings
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.simple\_simpa module
+---------------------------
+
+.. automodule:: urclib.simple_simpa
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.urc\_common module
+-------------------------
+
+.. automodule:: urclib.urc_common
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.urc\_fl module
+---------------------
+
+.. automodule:: urclib.urc_fl
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: urclib
+   :members:
+   :undoc-members:
+   :show-inheritance:
