+++ user_doc/source/api/urclib.fuzzylogic.rst
@@ -1,0 +1,72 @@
+urclib.fuzzylogic package
+=========================
+
+The `fuzzylogic` package is derived from the `Spatially Integrated Multivariate Probabilistic Assessment (SIMPA)
+<https://edx.netl.doe.gov/dataset/simpa-tool>`_.
+
+Submodules
+----------
+
+urclib.fuzzylogic.combiner module
+---------------------------------
+
+.. automodule:: urclib.fuzzylogic.combiner
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.fuzzylogic.embedgen module
+---------------------------------
+
+.. automodule:: urclib.fuzzylogic.embedgen
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.fuzzylogic.fuzzycurves module
+------------------------------------
+
+.. automodule:: urclib.fuzzylogic.fuzzycurves
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.fuzzylogic.fuzzylogic module
+-----------------------------------
+
+.. automodule:: urclib.fuzzylogic.fuzzylogic
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.fuzzylogic.geomutils module
+----------------------------------
+
+.. automodule:: urclib.fuzzylogic.geomutils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.fuzzylogic.nodata\_handling module
+-----------------------------------------
+
+.. automodule:: urclib.fuzzylogic.nodata_handling
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.fuzzylogic.settings module
+---------------------------------
+
+.. automodule:: urclib.fuzzylogic.settings
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: urclib.fuzzylogic
+   :members:
+   :undoc-members:
+   :show-inheritance:
