+++ user_doc/source/api/urclib.ui_qt.visualizer.qt_support.rst
@@ -1,0 +1,54 @@
+urclib.ui\_qt.visualizer.qt\_support package
+============================================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   urclib.ui_qt.visualizer.qt_support.pyqt5_support
+   urclib.ui_qt.visualizer.qt_support.pyside6_support
+
+Submodules
+----------
+
+urclib.ui\_qt.visualizer.qt\_support.GradientDlg module
+-------------------------------------------------------
+
+.. automodule:: urclib.ui_qt.visualizer.qt_support.GradientDlg
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.ui\_qt.visualizer.qt\_support.colorbuttons\_qt module
+------------------------------------------------------------
+
+.. automodule:: urclib.ui_qt.visualizer.qt_support.colorbuttons_qt
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.ui\_qt.visualizer.qt\_support.qt\_GaiaGLWidget module
+------------------------------------------------------------
+
+.. automodule:: urclib.ui_qt.visualizer.qt_support.qt_GaiaGLWidget
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.ui\_qt.visualizer.qt\_support.qt\_color\_utils module
+------------------------------------------------------------
+
+.. automodule:: urclib.ui_qt.visualizer.qt_support.qt_color_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: urclib.ui_qt.visualizer.qt_support
+   :members:
+   :undoc-members:
+   :show-inheritance:
