+++ user_doc/source/api/urclib.ui_qt.rst
@@ -1,0 +1,72 @@
+urclib.ui\_qt package
+=====================
+
+The `ui_qt` package contains the Graphical User Interface GUI used by the URC Assessment Method tool. While this package
+may not be directly helpful in other projects, it can serve as a useful reference.
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   urclib.ui_qt.visualizer
+
+Submodules
+----------
+
+urclib.ui\_qt.about\_dlg module
+-------------------------------
+
+.. automodule:: urclib.ui_qt.about_dlg
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.ui\_qt.elide\_label\_widget module
+-----------------------------------------
+
+.. automodule:: urclib.ui_qt.elide_label_widget
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.ui\_qt.prog\_log\_dlg module
+-----------------------------------
+
+.. automodule:: urclib.ui_qt.prog_log_dlg
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.ui\_qt.results\_dlg module
+---------------------------------
+
+.. automodule:: urclib.ui_qt.results_dlg
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.ui\_qt.unified\_window module
+------------------------------------
+
+.. automodule:: urclib.ui_qt.unified_window
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.ui\_qt.view\_models module
+---------------------------------
+
+.. automodule:: urclib.ui_qt.view_models
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: urclib.ui_qt
+   :members:
+   :undoc-members:
+   :show-inheritance:
