+++ user_doc/source/api/urclib.ui_qt.visualizer.rst
@@ -1,0 +1,64 @@
+urclib.ui\_qt.visualizer package
+================================
+
+This visualization package is derived from an internal project used across several software tools developed by NETL-RIC.
+It is intended for showing GIS/map data in an interactive fashion while requiring few external dependencies.
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   urclib.ui_qt.visualizer.qt_support
+
+Submodules
+----------
+
+urclib.ui\_qt.visualizer.LayerCaching module
+--------------------------------------------
+
+.. automodule:: urclib.ui_qt.visualizer.LayerCaching
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.ui\_qt.visualizer.geometryglscene module
+-----------------------------------------------
+
+.. automodule:: urclib.ui_qt.visualizer.geometryglscene
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.ui\_qt.visualizer.ogrglscene module
+------------------------------------------
+
+.. automodule:: urclib.ui_qt.visualizer.ogrglscene
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.ui\_qt.visualizer.shaders module
+---------------------------------------
+
+.. automodule:: urclib.ui_qt.visualizer.shaders
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+urclib.ui\_qt.visualizer.textrenderer module
+--------------------------------------------
+
+.. automodule:: urclib.ui_qt.visualizer.textrenderer
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: urclib.ui_qt.visualizer
+   :members:
+   :undoc-members:
+   :show-inheritance:
