+++ .github/workflows/python-app.yml
@@ -33,7 +33,7 @@ jobs:
         python3 -m pip install pip==23.0.1
         python3 -m pip install flake8 pytest
         # install gdal here explicitly to match version with lib
-        python3 -m pip install -r .github/workflows/requirements.txt
+        python3 -m pip install -r requirements.txt
         python3 -m pip install gdal==$(gdal-config --version)	
     # - name: Lint with flake8
     #  run: |

[…]
