### 2023-06-20 #5500 @elbeejay
@justaPCWingo please update us with your progress, and provide a timeline for when you might expect to get those revisions taken care of so that our reviewers will have an idea of when they'll need to put this item back on their to-do lists.

Re: the geospatial/GDAL dependency question, my 2-cents are to use `conda` as `pip` is less capable of resolving dependency conflicts in non-python codebases (e.g., the underlying C/C++ powering GDAL. Another option to improve reproducibility of the environment could be creating some sort of container (e.g., using Docker) with the necessary libraries.
