### 2023-06-20 #5500 @justaPCWingo
@elbeejay I'm shooting to have addressed existing issues and to provide review responses by Monday, June 26th. I believe that is doable with my schedule and timeline. I have already begun making changes based on the feedback from @FrancescoPerrone and @jeinsle . If it looks like that date will slip, I'll post an update here.

@elbeejay unfortunately, docker is essentially a nonstarter here; our cyber policy keeps the command set for such virtualization solutions disabled (it blocks the use of Microsoft's WSL as well, which would be useful for cross-platform development). Generally, I use `conda` on windows precisely for the reason you mentioned: binary DLLs underpinning so many of the libraries. When on a Linux or Mac system, I don't find it as necessary and generally stick with `pip` for simplicity.

Thanks!
