### 2023-06-20 84b4086
updated readme with version numbers
