+++ paper.md
@@ -52,13 +52,13 @@ conventional CM deposits; such unconventional sources include _in situ_ geologic
 extraction [@osti_1891489].
  
 The extraction and recovery of conventional CM is a complex process traditionally involving strip mining, which is both 
-expensive and environmentally destructive [@BALARAM20191285]. Recent research has revealed that coralliferous sediments 
-may act as unconventional CM sources containing Rare-Earth Elements (REE) in significant concentrations [@SEREDIN201267];
-determining the likelihood and location of these resources in sedimentary basins, however, is both complex and 
-challenging. To address this, a new method of evaluating the potential occurrence of URC resources using a series of 
-validated heuristics has been developed [@CREASON2023]. While the entire process can be carried out manually using a 
-collection of tools, a new, standalone software tool has been developed to streamline and expedite the process: NETL's 
-`URC Resource Assessment Tool`.
+expensive and environmentally destructive [@BALARAM20191285]. Recent research has revealed that coralliferous
+may act as unconventional CM sources containing Rare-Earth Elements (REE) in significant concentrations 
+[@SEREDIN201267]; determining the likelihood and location of these resources in sedimentary basins, however, is both 
+complex and challenging. To address this, a new method of evaluating the potential occurrence of URC resources using 
+a series of validated heuristics has been developed [@CREASON2023]. While the entire process can be carried out 
+manually using a collection of tools, a new, standalone software tool has been developed to streamline and expedite 
+the process: NETL's `URC Resource Assessment Tool`.
 
 
 # Statement of need
@@ -97,30 +97,32 @@ of an analysis can be previewed within the tool itself (\autoref{fig:urc_out}).
 
 In addition to several core Python libraries, The following 3rd-party libraries were used to create this tool:
 
-* [GDAL](https://www.gdal.org): Used for general spatial data management and calculations [@gdal2022].
-* [NumPy](https://numpy.org/): Handled vector math and general numeric array management [@harris2020array]. 
-* [Pandas](https://pandas.pydata.org/): Utilized for statistical calculations for the _Data Available_ (DA) scoring 
-  [@reback2020pandas].
-* [SIMPA](https://edx.netl.doe.gov/dataset/simpa-tool): Tool for processing spatially explicit raster data using fuzzy 
-  logic statements. Used in _Data Supporting_ (DS) scoring [@simpa2019].
-* [PyQt](https://riverbankcomputing.com/software/pyqt/): Framework used to build the graphical user interface (GUI) for
-  the tool; built on [Qt](https://www.qt.io) [@pyqt2022; @qt2022].
-* [PyOpenGL](https://pyopengl.sourceforge.net/): Wrapper for [OpenGL](https://www.opengl.org/) API; used for map 
-  visualization [@pyOGL2022; @openGL2017].
-  * PyOpenGL-accelerate: Optional library which can increase the performance of PyOpenGL [@pyOGL2022].
-* [pyGLM](https://github.com/Zuzu-Typ/PyGLM): Python port of the [GLM](https://glm.g-truc.net/0.9.9/) library; used for
-  graphic-specific mathematics [@pyglm2022;@glm2020].
+* [GDAL](https://www.gdal.org), _v3.1.4_: Used for general spatial data management and calculations [@gdal2022].
+* [NumPy](https://numpy.org/), _v1.23.2_: Handled vector math and general numeric array management [@harris2020array]. 
[…]
