### 2023-06-21 babefce
updates and tweaks to source code documentation
