+++ user_doc/source/api/urclib.fuzzylogic.geomutils.rst
@@ -1,0 +1,45 @@
+urclib.fuzzylogic.geomutils
+===========================
+
+.. automodule:: urclib.fuzzylogic.geomutils
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      bounds_check
+      get_segmentlabelmappings
+      pretty_segment_names
+      pretty_segment_to_type
+      segment_types
+      to_pt2d
+      type_segment_to_pretty
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      BaseSegment
+      BezierSegment
+      LinearSegment
+      Pt2D
+      StepwiseSegment
+   
+   
+
+   
+   
+   
+
+
+
