+++ user_doc/source/api/urclib.ui_qt.visualizer.textrenderer.rst
@@ -1,0 +1,31 @@
+urclib.ui\_qt.visualizer.textrenderer
+=====================================
+
+.. automodule:: urclib.ui_qt.visualizer.textrenderer
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      AtlasEntry
+      StringEntry
+      TxtRenderer
+   
+   
+
+   
+   
+   
+
+
+
