+++ user_doc/source/api/urclib.fuzzylogic.combiner.rst
@@ -1,0 +1,29 @@
+urclib.fuzzylogic.combiner
+==========================
+
+.. automodule:: urclib.fuzzylogic.combiner
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      FLCombiner
+   
+   
+
+   
+   
+   
+
+
+
