+++ user_doc/source/api/urclib.fuzzylogic.embedgen.rst
@@ -1,0 +1,29 @@
+urclib.fuzzylogic.embedgen
+==========================
+
+.. automodule:: urclib.fuzzylogic.embedgen
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      generate_embeddable
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
