+++ user_doc/source/api/urclib.ui_qt.view_models.rst
@@ -1,0 +1,30 @@
+urclib.ui\_qt.view\_models
+==========================
+
+.. automodule:: urclib.ui_qt.view_models
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      ResultTreeModel
+      TaskListMdl
+   
+   
+
+   
+   
+   
+
+
+

[…]
