### 2023-06-21 ef2a389
updates to fix build/hosting with readthedocs.
