+++ .readthedocs.yaml
@@ -1,4 +1,4 @@
-version: 1
+version: 2
 
 build:
   os: "ubuntu-20.04"
@@ -6,6 +6,7 @@ build:
     python: "3.8"
 
 sphinx:
+  configuration: user_doc/conf.py
   fail_on_warning: true
   
 python:
