### 2023-06-21 c783de2
undo bad part of new rtd config
