+++ .readthedocs.yaml
@@ -6,7 +6,6 @@ build:
     python: "3.8"
 
 sphinx:
-  configuration: user_doc/conf.py
   fail_on_warning: true
   
 python:
