### 2023-06-21 9c2d2ce
change to urclib imports to allow version number to be imported into conf without importing everything else.
