### 2023-06-21 ddfe0f9
hardcoding version number to see if it fixes rtd build
