+++ user_doc/source/conf.py
@@ -14,8 +14,9 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath('../..'))
 
-from urclib._version import __version__ as urc_version
-
+# from urclib._version import __version__ as urc_version
+# for document testing
+urc_version='1.0'
 # -- Project information -----------------------------------------------------
 
 project = 'URC Resource Assessment Method Tool Documentation'
