### 2023-06-21 12c15aa
updated requirements for RTD YAML
