+++ .readthedocs.yaml
@@ -11,3 +11,4 @@ sphinx:
 python:
   install:
     - requirements: user_doc/sphinx_requirements.txt
+    - requirements: requirements.txt
