### 2023-06-21 6625659
attempt to use autoapi to fix build issues with RTD
