+++ user_doc/source/api/urclib.ui_qt.results_dlg.rst
@@ -1,29 +1,0 @@
-urclib.ui\_qt.results\_dlg
-==========================
-
-.. automodule:: urclib.ui_qt.results_dlg
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      ResultDlg
-   
-   
-
-   
-   
-   
-
-
-
