+++ user_doc/source/api/urclib.ui_qt.about_dlg.rst
@@ -1,29 +1,0 @@
-urclib.ui\_qt.about\_dlg
-========================
-
-.. automodule:: urclib.ui_qt.about_dlg
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      AboutDlg
-   
-   
-
-   
-   
-   
-
-
-
