+++ user_doc/source/api/urclib.ui_qt.visualizer.textrenderer.rst
@@ -1,31 +1,0 @@
-urclib.ui\_qt.visualizer.textrenderer
-=====================================
-
-.. automodule:: urclib.ui_qt.visualizer.textrenderer
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      AtlasEntry
-      StringEntry
-      TxtRenderer
-   
-   
-
-   
-   
-   
-
-
-
