+++ user_doc/source/api/urclib.ui_qt.visualizer.qt_support.pyqt5_support.rst
@@ -1,31 +1,0 @@
-urclib.ui\_qt.visualizer.qt\_support.pyqt5\_support
-===================================================
-
-.. automodule:: urclib.ui_qt.visualizer.qt_support.pyqt5_support
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-
-.. rubric:: Modules
-
-.. autosummary::
-   :toctree:
-   :recursive:
-
-   urclib.ui_qt.visualizer.qt_support.pyqt5_support.ui_gradientdlg
-
