+++ user_doc/source/api/urclib.ui_qt.elide_label_widget.rst
@@ -1,29 +1,0 @@
-urclib.ui\_qt.elide\_label\_widget
-==================================
-
-.. automodule:: urclib.ui_qt.elide_label_widget
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      ElideLabel
-   
-   
-
-   
-   
-   
-
-
-
