+++ user_doc/source/api/urclib.da_calc.rst
@@ -1,30 +1,0 @@
-urclib.da\_calc
-===============
-
-.. automodule:: urclib.da_calc
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      calc_sum
-      run_pe_score_da
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-
