+++ user_doc/source/api/urclib.fuzzylogic.rst
@@ -1,37 +1,0 @@
-﻿urclib.fuzzylogic
-=================
-
-.. automodule:: urclib.fuzzylogic
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-
-.. rubric:: Modules
-
-.. autosummary::
-   :toctree:
-   :recursive:
-
-   urclib.fuzzylogic.combiner
-   urclib.fuzzylogic.embedgen
-   urclib.fuzzylogic.fuzzycurves
-   urclib.fuzzylogic.fuzzylogic
-   urclib.fuzzylogic.geomutils
-   urclib.fuzzylogic.nodata_handling
-   urclib.fuzzylogic.settings
-
