+++ user_doc/source/api/urclib.urc_common.rst
@@ -1,50 +1,0 @@
-urclib.urc\_common
-==================
-
-.. automodule:: urclib.urc_common
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      build_pandas_dataframe
-      combine_domdist_rasters
-      dataframe_to_rastergroup
-      find_domain_component_rasters
-      find_unique_components
-      gen_domain_hitmaps
-      gen_domain_index_rasters
-      list_featureclass_names
-      list_featureclasses
-      mult_band_data
-      norm_lg_rasters
-      norm_multrasters
-      normalize_raster
-      raster_copy
-      raster_distance
-      rasterize_components
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      RasterGroup
-   
-   
-
-   
-   
-   
-
-
-
