+++ user_doc/source/api/urclib.ui_qt.visualizer.shaders.rst
@@ -1,38 +1,0 @@
-urclib.ui\_qt.visualizer.shaders
-================================
-
-.. automodule:: urclib.ui_qt.visualizer.shaders
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      ExportShadersToFiles
-      assignUniformBlock
-      buildShaders
-      findUniformLocations
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      ShaderProgMgr
-   
-   
-
-   
-   
-   
-
-
-
