+++ user_doc/source/api/urclib.ui_qt.visualizer.geometryglscene.rst
@@ -1,43 +1,0 @@
-urclib.ui\_qt.visualizer.geometryglscene
-========================================
-
-.. automodule:: urclib.ui_qt.visualizer.geometryglscene
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      dummyFn
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      GeometryGLScene
-      POLY_GRAD_IND
-   
-   
-
-   
-   
-   .. rubric:: Exceptions
-
-   .. autosummary::
-   
-      GaiaGLException
-      GaiaGLShaderException
-   
-   
-
-
-
