+++ user_doc/source/api/urclib.ui_qt.visualizer.qt_support.rst
@@ -1,36 +1,0 @@
-urclib.ui\_qt.visualizer.qt\_support
-====================================
-
-.. automodule:: urclib.ui_qt.visualizer.qt_support
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-
-.. rubric:: Modules
-
-.. autosummary::
-   :toctree:
-   :recursive:
-
-   urclib.ui_qt.visualizer.qt_support.GradientDlg
-   urclib.ui_qt.visualizer.qt_support.colorbuttons_qt
-   urclib.ui_qt.visualizer.qt_support.pyqt5_support
-   urclib.ui_qt.visualizer.qt_support.pyside6_support
-   urclib.ui_qt.visualizer.qt_support.qt_GaiaGLWidget
-   urclib.ui_qt.visualizer.qt_support.qt_color_utils
-
