+++ user_doc/source/index.rst
@@ -21,7 +21,5 @@ implements several geospatial analyses as part of the URC-resource assessment me
    appendices/index
    references
 
-   api/index
-
 
 .. include:: ../../DISCLAIMER.md
