### 2023-06-22 d2fe213
first stab at reviewer comments
