+++ paper.md
@@ -44,17 +44,16 @@ bibliography: paper.bib
 # Summary
 
 [U]{.underline}nconventional [R]{.underline}are-earth elements & [C]{.underline}ritical minerals (URC) [@osti_1891489] 
-are crucial to a growing number of industries worldwide [@BALARAM20191285]. Critical Minerals (CM) are minerals used in
-manufacturing which are essential to economic and national security while being vulnerable to supply disruption through 
-any number of external factors [@osti_1891489]. _Unconventional_ CM resources contrast with conventional CM resources 
-in that they are sourced from geologic or byproduct hosts distinctly separate from the mechanisms which establish 
-conventional CM deposits; such unconventional sources include _in situ_ geologic deposits and byproducts of industrial 
-extraction [@osti_1891489].
+are crucial to a growing number of industries worldwide [@BALARAM20191285]. Due to their use in manufacturing, Critical
+Minerals (CM) are essential to economic and national security, yet have supply chains vulnerable to external 
+disruptions[@osti_1891489]. _Unconventional_ CM are sourced from geologic or byproduct hosts distinctly separate from 
+the mechanisms which establish conventional CM deposits [@osti_1891489]. Unconventional sources for CM include 
+_in situ_ geologic deposits and byproducts of industrial extraction [@osti_1891489].
  
 The extraction and recovery of conventional CM is a complex process traditionally involving strip mining, which is both 
-expensive and environmentally destructive [@BALARAM20191285]. Recent research has revealed that coralliferous
+expensive and environmentally destructive [@BALARAM20191285]. Recent research has revealed that coaliferous sediments
 may act as unconventional CM sources containing Rare-Earth Elements (REE) in significant concentrations 
-[@SEREDIN201267]; determining the likelihood and location of these resources in sedimentary basins, however, is both 
+[@SEREDIN201267]. Determining the likelihood and location of REE resources in sedimentary basins is both 
 complex and challenging. To address this, a new method of evaluating the potential occurrence of URC resources using 
 a series of validated heuristics has been developed [@CREASON2023]. While the entire process can be carried out 
 manually using a collection of tools, a new, standalone software tool has been developed to streamline and expedite 
@@ -79,11 +78,10 @@ From this point, a Data Availability (DA) and / or a Data Supporting (DS) analys
 (\autoref{fig:pes_flow}); both analyses will operate on a vector-based spatial dataset describing the target formation,
 following the labelling scheme specified in the supplementary material in @CREASON2023. These data are rasterized 
 according to the grid specification of the aforementioned domains with each cell tagged with the appropriate set of 
-indices. In the case of a DA analysis, each pixel in the rasterized data is evaluated according to \autoref{eq:da} as 
-described by @CREASON2023, producing a DA score for each cell that is unique to each geologic resource type. For the DS 
-analysis, Spatial Implicit Multivariate Probability Assessment method [@simpa2019] is applied using a series of 
-predefined Fuzzy Logic statements, which encompass the application of equations \autoref{eq:ds}, \autoref{eq:dsm}, and
-\autoref{eq:pe}. 
+indices. In the case of a DA analysis, each pixel in the rasterized data is evaluated by applying Equation (1) as 
+described in @CREASON2023, producing a DA score for each cell that is unique to each geologic resource type. For the DS 
+analysis, the Spatial Implicit Multivariate Probability Assessment (SIMPA) method [@simpa2019] is applied using a series
[…]
