+++ paper.bib
@@ -204,3 +204,10 @@ place = {United States},
 year = {2022},
 month = {10}
 }
+
+@misc{npsimd,
+author = {NumPy Developers },
+title = {CPU/SIMD Optimizations - Numpy v2.0dev0 Manual },
+publisher = { NumPy Developers },
+url = {https://numpy.org/devdocs/reference/simd/index.html}
+}
