+++ paper.bib
@@ -206,7 +206,7 @@ month = {10}
 }
 
 @misc{npsimd,
-author = {NumPy Developers },
+author = {{NumPy Developers}},
 title = {CPU/SIMD Optimizations - Numpy v2.0dev0 Manual },
 publisher = { NumPy Developers },
 url = {https://numpy.org/devdocs/reference/simd/index.html}
