### 2023-06-29 #5500 @justaPCWingo
Hi @elbeejay , @FrancescoPerrone , and @jameshgrn , I've posted a new draft of `paper.md` to the [`joss-submit` branch of the repository](https://github.com/NETL-RIC/URC-Assessment-Method/tree/joss-submit). I've also uploaded [a response to @FrancescoPerrone](https://github.com/NETL-RIC/URC-Assessment-Method/blob/joss-submit/response_to_FPerrone.pdf) and [a response to @jameshgrn](https://github.com/NETL-RIC/URC-Assessment-Method/blob/joss-submit/response_to_JGearon.pdf) to the repo with my answers in red text.

Thanks!
