### 2023-07-12 #5500 @jeinsle
Howdy. I am at a conference this week. So will dig into it next week

Sent from Outlook for iOS<https://aka.ms/o0ukef>
