### 2023-07-25 #5500 @jameshgrn
cheers to the authors for a well-done revision, pretty much everything looks good to me. I cant test it right now, but the paper as well as the new readme look good. a couple things now:

- Link to CONTRIBUTING and LICENSE in the original README.md
- add a sentence to check this box: "State of the field: Do the authors describe how this software compares to other commonly-used packages?"
- include installation guidelines in the readme

will follow up soon
