### 2023-07-31 #5500 @jeinsle
Echoing both @jameshgrn<https://github.com/jameshgrn>, and @FrancescoPerrone<https://github.com/FrancescoPerrone> I have been slammed with deadlines since I returned from holidays.  I have made some progress but need to write things down / comment on Git.

j

----------------------------------------------------------------------------------------------------
Sometimes I work outside of standard hours, but I do not expect replies during these times.

Joshua F. Einsle
Lord Kelvin Adam Smith Research Fellow in Data Science
School of Geographical and Earth Sciences
University of Glasgow

https://www.gla.ac.uk/schools/ges/staff/joshuafranzeinsle/
