### 2023-08-01 314198f
updated readme with links to License and contributing instructions. Added small section on installation
