### 2023-08-02 4f5040a
added some updates to paper to respond to review requests
