### 2023-08-07 #5500 @FrancescoPerrone
Dear all,

I appreciate your prompt and detailed responses to my initial review comments. I have reviewed the modifications made in response to my initial feedback, and here are my follow-up comments:

1. **Authors' contributions**: Thank you for detailing the contributions made by each author in your response. While this information is not strictly required in the JOSS submission documentation, it adds value to the review process and helps ensure that each individual's contributions are appropriately credited.

2. **Manuscript detail**: I am satisfied with the addition of the "Implementation Details" section in your manuscript, and the expansion of the example tutorial in your documentation. This does improve the reproducibility of the methods employed by the URC Resource Assessment Tool, and I appreciate your attention to this point.

3. **License file**: The renaming of your license file to "LICENSE" aligns with best practices for open-source software projects and enhances the clarity of your project's legal framework.

4. **Version information**: Thank you for including version information in the README, user documentation, and making it available through the `__version__` variable in `urclib`. This is a positive addition that aids in reproducibility and usage of your software.

5. **Testing instructions**: I appreciate your adding instructions for running tests using pytest to the documentation. This addition is helpful for users and contributors who wish to verify the software's functionality.

6. **Community Guidelines**: The inclusion of a CONTRIBUTING.md file in your repository enhances your project's openness to contributions, which is a critical aspect of open-source software.

Regarding the manuscript:

1. **Title**: While it may be a bit long, I understand your concern about introducing undefined acronyms. Retaining the current title is acceptable given this consideration.

2. **Summary**: The breaking down of sentences has indeed improved readability. Thank you for your attention to this detail.

3. **Methodology**: I understand your concerns about reproducing large amounts of detail from the Creason 2023 paper. Directing the reader to this paper for additional information is a reasonable solution.

4. **Support Libraries**: Thank you for specifying the versions of the libraries used. This enhances the reproducibility of your software.

On the basis of these modifications, I am satisfied that you have addressed all the major points raised in my initial review. I appreciate the constructive manner in which you engaged with my comments, and your commitment to improving both your manuscript and your software is clear.

**Recommendation**: Based on these changes, I recommend that the manuscript and software are accepted for publication in the Journal of Open Source Software (JOSS), pending any final checks by the editor.

Once again, congratulations on developing this valuable tool and for producing a clear and informative paper to accompany it.

Best regards,

Francesco

**Additional comments**

While I trust that you have made the changes you mentioned, I unfortunately was not able to run the tool again due to time constraints. However, I would like to provide the following general recommendations to further enhance the tool and its documentation:

- **Continuous Integration (CI)**: If it is not already in place, consider adding a CI workflow to automatically run tests when changes are pushed to the repository. This would add an additional layer of assurance that the software is working as intended.
- **Software Tutorial**: If time permits, consider developing a video tutorial or webinar for users. This can be especially helpful for complex software like yours, as it can help users visualize the process and better understand how to use the tool.
- **Use Cases**: Continue to add real-world use cases or examples to the documentation as they become available. These can serve as valuable guides for users to understand how the tool can be applied to their specific needs.
- **Active Community Engagement**: Aim to maintain an active presence in the repository to address questions, issues, or suggestions that users may post. This will encourage users to contribute and engage more with your project.

Again, these are suggestions for potential improvements and may be considered as the project evolves and grows.

On the basis of your responses and modifications, I am satisfied that you have addressed all the major points raised in my initial review. I appreciate the constructive manner in which you engaged with my comments, and your commitment to improving both your manuscript and your software is clear.

**Recommendation**: Given my inability to run the tool a second time due to time constraints, my recommendation is made based on the changes as stated by the authors. I recommend that the manuscript and software are accepted for publication in the Journal of Open Source Software (JOSS), pending any final checks by the editor and, if possible, a final verification by someone able to run the tool.

Once again, congratulations on developing this valuable tool and for producing a clear and informative paper to accompany it.
