### 2023-08-07 #5500 @elbeejay
Fantastic, thanks @FrancescoPerrone for the clear review summary. As it looks like you are satisfied with the revisions made, I'd just like to ask that you return to your reviewer checklist above and check off the outstanding items.

Thanks again for your efforts reviewing this submission to JOSS!
