### 2023-08-08 #5500 @jeinsle
All – here is my first pass at the review. I am breaking it down by the sections of the checklist

General Checks

Contribution – poking at the github, it looks like most of the work has been done by @justaPCWingo. That said I suspect that this is more to do with how the code was developed and released. - overall, it is hard for me to assess how much work was done by any one team member.  From my knowledge of the group the contributors on the Git and the authors make sense.

Functionality

Installation – as per my comment on the installation instructions (below) I have not been able to fully install. I know my MSc student has gotten this to run, so I am suspecting a machine problem.

Functionality – personally still need to test, my student is getting results.

Performance – N/A??

Documentation:

Statement of need – the statement of need in the JOSS manuscript clearly lays out who the audience is. It would benefit the ReadTheDoc introduction to add these sentences to the project description. I would also think that this would help to be placed in the Read me on the Github as it would just tell someone landing on the Github without either manuscript or the ReadTheDocs what the repository is about. It may seem redundant, but this reflects the fact the users will be coming to code at about four different directions and just trying to make a consistent experience.

Installation instructions – on the github I would push to make these slightly more explicit / expanded.  I have been trying to install on my system (Ubuntu 20.04.6) and have been unsuccessful.  In particular, I am hung up on installing GDAL. If this is meant to only be run on windows machines, please add that to the specification in the installation directions. My recommendation would be for a more step by step set along lines where you tell the user to:

Create an environment

Specify if to use Conda or pip for each package and what order they should be installed in.

My attempt has been using conda, but if I should be taking a different approach this needs to be explicit.

Example usage – is provided both in the original Creason et al manuscript as well as the ReadTheDocs file where the example data is worked through.

Functionality documentation – an autogenerated API is provided. This is proably fine for initial deployment, but contributions by users will likely change what is here.

Automated tests – yes standard auto testing has been set up

Community guidelines – yes, this is fairly straightforward and what would be expected for a project that has just been launched.

Software paper

Summary – yes the summary is clear and points out that this tool does one very specific job. (which is difficult and would be hard to replicate with out)

Statement of need – this is well written and clearly explains who and why this code is needed.

State of the field – it is not clear in this manuscript if there are other tools out there trying for similar approaches. It would help to know if there are other methods but that might be discussed in Creason et al (I need to re-read that again)

Quality of writing –overall this is clearly written document which explains the purpose of the code.

References -  again this section is fine.
