### 2023-08-08 #5500 @jeinsle
I have a few more things to test and get the GDAL running but that is a task for tomorrow
