### 2023-08-09 #5500 @justaPCWingo
Hi @FrancescoPerrone ,
   Thanks for the recommendation and additional comments. I will say that we have some basic CI implemented as part of the github repository, but I can easily see it being fleshed out more as this project continues to evolve. The suggestions of video tutorials, example use cases, and community engagement are solid suggestions, and will look for opportunities to implement these as time and budget permits!
