### 2023-08-09 #5500 @justaPCWingo
Hi @jeinsle ,
I like the suggestion of adding the paper's Statement of need to the readme and user documentation; would this raise concerns of self-plagiarism? Perhaps I'm overthinking it. If not, it would be an easy thing to do.

To answer some of the installation questions, this software was mostly developed under Windows, but there are no platform-specific logic or packages in any of this source code. The CI action used to run pytest is running the Python code under Ubuntu 20.x or 22.x (don't remember which exactly). The environment statement was intentionally left vague since how package installation is handled is largely up to personal preference. The packages required should work whether installed under a conda, Python venv, or raw environment. Package order also shouldn't matter.

There are a pair of files in the root of the repository that may assist with virtual environment building, but your mileage may vary.
* For conda: `conda env create -n URC --file environment.yml`.
* For venv: `python -mpip install -r requirements.txt`.

GDAL seems to be an incredible pain to install regardless of the platform, but its the best package to serve the needs of this project. Its hard to give specific instructions since GDAL is so dependent on dynamic libraries (.dll/.so/.dylib) that the issues that can crop up will likely be specific to each installation.

One thing I will do is walk through the setup on my personal Ubuntu machine and let you know what sort of bumps I run into. If we're lucky I'll hit the same pain point and be able to help you out.

RE: State of the field, I did revise the language a bit in response to @jameshgrn by essentially just adding this sentence:

> No other Python packages are known to contain the combination of geospatial information systems (GIS) and fuzzy logic support required to directly implement the method defined in Creason et al.

Would this be sufficient to address your concerns?
