### 2023-08-09 #5500 @justaPCWingo
Hi @jeinsle ,
For GDAL to work, we need to ensure that the library and the python bindings match. We can do that with the following steps (assume all commands are run from the root directory of the repository in a bash prompt):

1. Install GDAL library using `apt`:

```sh
sudo apt update
sudo apt install libgdal-dev gdal-data
```

2. Create Python virtual environment:

```sh
python3 -mvenv urc_env
source urc_env/bin/activate
```

3. Install gdal prior to the other packages, ensuring the version matches the  currently installed GDAL  library.

```sh
pip install gdal==$(gdal-config --version)
```

4. Install other dependencies using requirements.txt:

```sh
pip install -r requirements.txt
```

From here, the command `python ./urc_assessment_method.py` should launch the program.
