### 2023-08-11 #5500 @justaPCWingo
Hi @jeinsle ,
   RE: Self-plagiarism: fair enough. I'll look to incorporating the Statement of need's language into the Readme and userdocs.
Thanks for the the suggestion for language for providing installation advice. I'll give it a bit of thought on what kind of language and details to include, but will take a stab at providing some statement to that affect.

In my experience, I typically use `conda env` on Windows, and Python's `venv` on Linux. I've had better luck with Windows DLLs under conda, but find it too complicated when working on Linux. Again, your mileage may vary.
