### 2023-08-11 #5500 @elbeejay
Looks like a valuable discussion @jeinsle and @justaPCWingo, I'll weigh in on the two points briefly:

Re: Self-plagiarism - don't worry about self-plagiarism for this type of documentation, especially if you think adding that text into the readme and code documentation will improve the user experience.

Re: GDAL and installation - this is a common issue for folks doing geospatial work in Python, @justaPCWingo I think you could adopt some of the language from your comment above about the different installation pathways for GDAL and include it in the readme/documentation to help provide users with some guidance on that installation process as @jeinsle  suggests above. Alternatively I'm sure a link to some other content about installing GDAL for Python would serve the same goal.
