### 2023-08-15 3106278
Updated Readme and documentation with Statement of Need from journal pub and some advice on intalling venvs and GDAL
