+++ README.md
@@ -5,10 +5,48 @@ _Latest Release: 1.0.0_
 Logic for running **Unconventional Rare-Earth & Critical Minerals (URC)** analyses on structured data, wrapped into a
 tool.
 
-The tool documentation can be built using Sphinx with the content in the `user_doc` folder. Alternately, a compiled version of the 
-documentation is mirrored on ReadTheDocs.io [here](https://urc-assessment-method.readthedocs.io/en/latest/).
+The tool documentation can be built using Sphinx with the content in the `user_doc` folder. Alternately, a compiled 
+version of the documentation is mirrored on ReadTheDocs.io 
+[here](https://urc-assessment-method.readthedocs.io/en/latest/).
 
-The [license](./LICENSE) and [contribution instructions](./CONTRIBUTING.md) are also included as part of this distribution.
+The [license](./LICENSE) and [contribution instructions](./CONTRIBUTING.md) are also included as part of this 
+distribution.
+
+---
+## Statement of Need
+
+The **URC Resource Assessment Method** applies the data analysis methods outlined in [Creason et al.][1] This tool 
+is a complete application written in Python and built on top of several [open-source libraries](#package-dependencies). 
+No other Python packages are known to contain the combination of geospatial information systems (GIS) and fuzzy logic 
+support required to directly implement the method defined in [Creason et al.][1] The intended users for this tool are 
+geologists and geospatial scientists who are looking to better understand the mode and spatial distribution of potential
+URC resource occurrences in sedimentary basins.
+
+There are several ways that the **URC Resource Assessment Method** can be configured to run, but fundamentally the tool 
+takes in a collection of spatial domains which fall under the Lithological, Structural, and Secondary Alteration 
+categories defined by the [Subsurface Trend Analysis (STA) method][2]. These domains are combined, clipped to a 
+researcher-defined boundary, and grided to cells of a research-specified dimension.
+
+From this point, a Data Availability (DA) and / or a Data Supporting (DS) analysis can be undertaken by the tool; both 
+analyses will operate on a vector-based spatial dataset describing the target formation, following the labelling scheme 
+specified in the supplementary material in [Creason et. al.][1] These data are rasterized according to the grid 
+specification of the aforementioned domains with each cell tagged with the appropriate set of indices. In the case of a 
+DA analysis, each pixel in the rasterized data is evaluated by applying Equation (1) as described in 
+[Creason et. al.][1], producing a DA score for each cell that is unique to each geologic resource type. For the DS 
+analysis, the [Spatial Implicit Multivariate Probability Assessment (SIMPA) method][3] is applied using a series of 
+predefined Fuzzy Logic statements, encapsulating the logic of Equations (2), (3), and (4) in [Creason et. al.][1]
+
+The **URC Resource Assessment Method** can be run either using the standalone GUI, or as a command-line tool. The 
+former configuration is useful for a guided approach to executing the URC mineral-related analyses and previewing the 
+results in the tool itself, whereas the latter is useful for integration of the tool into a workflow as part of a batch
+process. Regardless of how it is run, the results of the requested analyses are written to GeoTIFF files, which can be
+imported into most GIS analysis tools. Optionally, when run from the GUI the results of an analysis can be previewed 
+within the tool itself.
+
+[1]: https://doi.org/10.1007/s11053-023-10163-x
+[2]: https://doi.org/10.1190/INT-2019-0019.1
+[3]: https://edx.netl.doe.gov/dataset/simpa-tool
+
 
 ---
 
@@ -16,6 +54,9 @@ The [license](./LICENSE) and [contribution instructions](./CONTRIBUTING.md) are 
 
[…]
