### 2023-08-15 #5500 @justaPCWingo
Hi all,
   I've updated the [readme](https://github.com/NETL-RIC/URC-Assessment-Method/blob/joss-submit/README.md) and [user documentation](https://urc-assessment-method.readthedocs.io/en/latest/index.html) with the suggestions provided. Please let me know if that is satisfactory.

Thanks!
