### 2023-08-29 #5500 @jeinsle
@justaPCWingo the good news is that I can get the URC to install and launch.  However, when I download the ```esm3.zip```  file I can not seem the find the ESRI shape files that the UI is looking for even after unzipping ```ecm_3/PRB- URC Asessment data/DA & DS databases/DA & DS vector database.zip```

just a little confussed
