### 2023-08-29 554f224
fix to some typos in example.md
