### 2023-08-29 #5500 @justaPCWingo
@jeinsle , the first problem I see is `ecm_3` is a typo; it should be `esm_3`. I just pushed changes to address these typos. which files are you having a hard time locating specifically?

I see the shp files `PRB_structural_domains.shp`  and  `PRB_lithologic_domains.shp` in the `esm_3/PRB- URC Asessment data/Domains/`. Note that this is in the outer zip file (that is the zip file as downloaded from EDX).

The `PRB_URC_DA_DS_Database.gdb` file is in the inner zip file (`esm_3/PRB- URC Asessment data/DA & DS databases/DA & DS vector database.zip`).

I don't know why the decision was made to stick a zip file inside another zip file, as I had no involvement in the packaging of this EDX submission. Is that where the confusion is coming from?
