### 2023-08-31 #5500 @jeinsle
@justaPCWingo sorry for delay as I do want to close this out sooner than later.

1. re: esm vs ecm... the ecm above is a typo for the post only.  I did download the esm3.zip file but for some reason I was not finding the 'Domains' directory the other day and so was not seeing the *.shp files - today found them and I am running the tutorial as we speak.  With that I am happy. I can run  the tutorial as directed and reprodice the results.  Also the documentation is much improved. overall I think I am satsfied and would think that this is all done now??? @elbeejay  what is left to finish off the review?
