### 2023-09-01 #5500 @elbeejay
@justaPCWingo - I'm glad to read positive final notes from all of the reviewers. At this time I'm going to do a final check of the repository and paper, and then I'll have a few final steps I will ask you to complete to square away and finalize the metadata that'll be associated with the JOSS publication.
