### 2023-09-01 #5500 @elbeejay
@justaPCWingo I've got to ask you to fix the error in the `.bib` file referenced above regarding the `osti_1891489` reference. You should be able to match the syntax of one of the other multi-author references to accomplish this. I'd also ask that you consider explicitly telling users they need to clone the repository locally in order to install it (as from what I can tell this is not installable via `pip` or `conda`, and instructions assume access to repository files).

After making revisions you can prompt the bot the way I did above to check the references section.

Once the references are fixed and that check is executed successfuly, I will need you to do the following:

- [ ] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) - for you this is just updating the archive to the latest versions
