### 2023-09-01 #5500 @justaPCWingo
@elbeejay strange, I'll take a look at the bib and see what's going on. And yes, the code is not presently hosted on pypi or a conda channel. I'll update the instructions as you suggest.

Would the [Energy Data Exchange (EDX)](https://edx.netl.doe.gov) qualify as an institutional repository for the purposes of the checklist above? This is where we typically park our public products.
