+++ paper.bib
@@ -197,9 +197,7 @@ title = {Unlocking the Potential of Unconventional Critical Mineral Resources St
 author = {Yesenchak, Rachel and Justman, Devin and Bauer, Sophia and Creason, C Gabriel and Gordon, Andrew and Montross, Scott N and Sabbatino, Michael and Rose, Kelly},
 abstractNote = {An ArcGIS Story Map that provides context and understanding of unconventional critical mineral resource potential.},
 doi = {10.18141/1891489},
-url = {https://www.osti.gov/biblio/1891489}, journal = {},
-number = ,
-volume = ,
+url = {https://edx.netl.doe.gov/dataset/unlocking-the-potential-of-unconventional-critical-mineral-resources-story-map},
 place = {United States},
 year = {2022},
 month = {10}
