### 2023-09-01 a3b0edd
added instructions for pulling repository as part of installation process. Locked versions of stuff for RTD since they managed to break their own stuff when updating their theme
