+++ user_doc/sphinx_requirements.txt
@@ -1,5 +1,5 @@
-sphinx
+sphinx==7.1.4
+sphinx_rtd_theme==1.2.2
 myst-parser
 sphinxcontrib-bibtex
-sphinx_rtd_theme
 sphinx-autoapi
