### 2023-09-01 267ef7b
another attempted fix for RTD
