+++ user_doc/sphinx_requirements.txt
@@ -1,5 +1,5 @@
-sphinx==7.1.4
-sphinx_rtd_theme==1.2.2
+sphinx
+sphinx_rtd_theme
 myst-parser
 sphinxcontrib-bibtex
 sphinx-autoapi
