### 2023-09-01 #5500 @justaPCWingo
@elbeejay : I've (hopefully) addressed the issues with the bib, and added a little bit of information to the readme with a sentence pointing out the repo will need to be downloaded as part of the installation process.

There is an existing entry on [EDX](https://edx.netl.doe.gov/) for a pyinstaller-bundled version of this tool, but it does not share the same name as the paper included in this entry. if the archive is supposed to be just the tagged version of the tool, source, I can either attach to the existing submission (which has a doi) or create a new submission altogether.

Which would be preferrable?
