### 2023-09-01 #5500 @justaPCWingo
@elbeejay one more question: should the tagged release be from the branch with the paper and bibliography, or the master branch which doesn't contain JOSS specific assets?
