### 2023-09-02 #5500 @elbeejay
@justaPCWingo you'll need to fix the invalid DOI issues flagged by the bot above.

Regarding the question about [EDX](https://edx.netl.doe.gov/), I think the issue is going to be that the archived code is not accessible without creation of an EDX account. It is also unclear to me whether or not EDX creates a DOI (does it?). @kthyng if you don't mind weighing in here that'd be great, but my thought is that we'd expect the archived versions of the code to follow the same [guidelines](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) as the code repositories.

@justaPCWingo we can wait on @kthyng to weigh in, but with respect to your other questions we ask that the institutional repository title and authors match the JOSS paper exactly, so in this case that may require a new entry. We want the tagged release to come from the primary development branch that holds the latest code updates reflecting the version of the tool the reviewers approved of.

Thanks
