### 2023-09-02 #5500 @justaPCWingo
@elbeejay Thanks for the info FYI, you _can_ download public resources from EDX, but you _can't_ preview the resources; I verified that this was the case. You can optionally create a DOI as part of the submission process on EDX.

Thanks to your responses, I think I know the best course of action here. Zenodo is the way to go in this case. I need to get permission to upload to Zenodo, but since this is just a snapshot of existing publically released code, I don't foresee an issue, just some CYA.

That's odd that the DOI prefix is causing issues; I haven't run into that with other *.bib collators. I'll fix and let you know when its up.
