### 2023-09-02 cc92dd4
fixed some DOI numbers on paper bib to fix bot errors with JOSS
