+++ paper.bib
@@ -89,7 +89,7 @@ pages = {67-93},
 year = {2012},
 note = {Minerals and Trace Elements in Coal},
 issn = {0166-5162},
-doi = {https://doi.org/10.1016/j.coal.2011.11.001},
+doi = {10.1016/j.coal.2011.11.001},
 url = {https://www.sciencedirect.com/science/article/pii/S0166516211002448},
 author = {Vladimir V. Seredin and Shifeng Dai},
 keywords = {Coal deposits, Rare earth elements and yttrium (REY), REY distribution, REY mode of occurrence},
@@ -104,7 +104,7 @@ number = {4},
 pages = {1285-1303},
 year = {2019},
 issn = {1674-9871},
-doi = {https://doi.org/10.1016/j.gsf.2018.12.005},
+doi = {10.1016/j.gsf.2018.12.005},
 url = {https://www.sciencedirect.com/science/article/pii/S1674987119300258},
 author = {V. Balaram},
 keywords = {REE, Mineralogy, Deposits, Metallurgy, Environmental, Recycling},
@@ -119,7 +119,7 @@ author   = {Creason, C.G., Justman, D., Rose, K., Montross, S., Bean, A., Mark-M
 year     = {2023},
 month    = {2},
 day      = {28},
-doi      = {https://doi.org/10.1007/s11053-023-10163-x},
+doi      = {10.1007/s11053-023-10163-x},
 keywords = {resource assessment, ree enrichment, knowledge-data framework, fuzzy logic},
 abstract = {Rare-earth elements (REE) supply raw materials that constitute many of our modern critical infrastructure, 
             defense, technology, and electrification needs. Despite REE accumulations occurring in conventional 
