### 2023-09-02 #5500 @justaPCWingo
@elbeejay I pushed fixes for the DOI errors flagged above. Definitely something I screwed up, which at least made it obvious to fix :)
