### 2023-09-03 #5500 @elbeejay
> Thanks to your responses, I think I know the best course of action here. Zenodo is the way to go in this case. I need to get permission to upload to Zenodo, but since this is just a snapshot of existing publically released code, I don't foresee an issue, just some CYA.

This sounds good, let me know when you've been able to do this @justaPCWingo - thanks for being understanding and flexible.
