### 2023-09-05 #5500 @justaPCWingo
Hi @elbeejay ,

the tagged version is v1.0.1.
the zenodo DOI is 10.5281/zenodo.8319843

Please let me know if there is anything else you need me to do.
