### 2023-09-06 #5500 @kthyng
Sounds like you all went with Zenodo, let me know if you still have questions.
